+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/playlists/PlaylistsServiceClient.java
@@ -28,6 +28,13 @@ public class PlaylistsServiceClient extends VPinStudioClientService {
     return list;    
   }
 
+
+  public List<PlaylistRepresentation> getPlaylistTree() {
+    PlaylistRepresentation[] playlists = getRestClient().get(API + "playlists/tree", PlaylistRepresentation[].class);
+    ArrayList<PlaylistRepresentation> list = new ArrayList<>(Arrays.asList(playlists));
+    return list;
+  }
+
   public PlaylistRepresentation getPlaylist(int playlistId) {
     return getRestClient().get(API + "playlists/" + playlistId, PlaylistRepresentation.class);
   }